Module iroh_net::discovery::pkarr

A discovery service which publishes and resolves node information to a pkarr relay.

Modules

dht

Pkarr based node discovery for iroh-net

Structs

PkarrPublisher

Publish node info to a pkarr relay.

PkarrRelayClient

A pkarr client to publish [pkarr::SignedPacket]s to a pkarr relay.

PkarrResolver

Resolve node info using a pkarr relay.

Constants

DEFAULT_PKARR_TTL

Default TTL for the records in the pkarr signed packet. TTL tells DNS caches how long to store a record. It is ignored by the iroh-dns-server as the home server keeps the records for the domain. When using the pkarr relay no DNS is involved and the setting is ignored.

DEFAULT_REPUBLISH_INTERVAL

Interval in which we will republish our node info even if unchanged: 5 minutes.

N0_DNS_PKARR_RELAY_PROD

The pkarr relay run by n0, for production.

N0_DNS_PKARR_RELAY_STAGING

The pkarr relay run by n0, for testing.