Module iroh_net::endpoint

The Endpoint allows establishing connections to other iroh-net nodes.

The Endpoint is the main API interface to manage a local iroh-net node. It allows connecting to and accepting connections from other nodes. See the module docs for more details on how iroh-net connections work.

The main items in this module are:

Endpoint to establish iroh-net connections with other nodes.

Builder to create an Endpoint.

Structs

Accept

Future produced by Endpoint::accept.

AddrInfo

Network paths to contact an iroh-net node.

Builder

Builder for Endpoint.

Connecting

In-progress connection attempt future

Connection

A QUIC connection.

ConnectionTypeStream

Stream returning ConnectionTypes

DirectAddr

A direct address on which an iroh-node might be contactable.

DirectAddrInfo

Information about a direct address.

DirectAddrsStream

Stream returning local endpoints as they change.

Endpoint

Controls an iroh-net node, establishing connections with other nodes.

NodeAddr

Network-level addressing information for an iroh-net node.

RecvStream

A stream that can only be used to receive data

RemoteInfo

Details about a remote iroh-net node which is known to this node.

SendStream

A stream that can only be used to send data

TransportConfig

Parameters governing the core QUIC state machine

VarInt

An integer less than 2^62

Enums

ConnectionError

Reasons why a connection might be lost

ConnectionType

The type of connection we have to the endpoint.

ControlMsg

The type of control message we have received.

DirectAddrType

The type of direct address.

ReadError

Errors that arise from reading from a stream.

WriteError

Errors that arise from writing to a stream

Functions

default_relay_mode

Returns the default relay mode.

get_remote_node_id

Extract the PublicKey from the peer’s TLS certificate.

make_server_config

Creates a [quinn::ServerConfig] with the given secret key and limits.