Module iroh_net::magicsock

Implements a socket that can change its communication path while in use, actively searching for the best way to communicate.

Based on tailscale/wgengine/magicsock

DEV_RELAY_ONLY env var:

When present at compile time, this env var will force all packets to be sent over the relay connection, regardless of whether or not we have a direct UDP address for the given node.

The intended use is for testing the relay protocol inside the MagicSock to ensure that we can rely on the relay to send packets when two nodes are unable to find direct UDP connections to each other.

This also prevent this node from attempting to hole punch and prevents it from responding to any hole punching attempts. This node will still, however, read any packets that come off the UDP sockets.

Re-exports

pub use self::metrics::Metrics;

pub use self::node_map::ConnectionType;

pub use self::node_map::ConnectionTypeStream;

pub use self::node_map::ControlMsg;

pub use self::node_map::DirectAddrInfo;

pub use self::node_map::NodeInfo as ConnectionInfo;

Modules

metrics 🔒

node_map 🔒

relay_actor 🔒

timer 🔒

udp_conn 🔒

Structs

Actor 🔒

DirectAddr

A direct address on which an iroh-node might be contactable.

DirectAddrUpdateState 🔒

Manages currently running direct addr discovery, aka netcheck runs.

DirectAddrsStream

Stream returning local endpoints as they change.

DiscoSecrets 🔒

DiscoveredDirectAddrs 🔒

The discovered direct addresses of this MagicSock.

Handle 🔒

Handle for MagicSock.

MagicSock 🔒

Iroh connectivity layer.

NetInfo 🔒

Contains information about the host’s network state.

Options 🔒

Contains options for MagicSock::listen.

PacketSplitIter 🔒

Splits a packet into its component items.

QuicMappedAddr 🔒

The fake address used by the QUIC layer to address a node.

Enums

ActorMessage 🔒

DirectAddrType

The type of direct address.

DiscoBoxError 🔒

DiscoMessageSource 🔒

Constants

ENDPOINTS_FRESH_ENOUGH_DURATION 🔒

How long we consider a STUN-derived endpoint valid for. UDP NAT mappings typically expire at 30 seconds, so this is a few seconds shy of that.

HEARTBEAT_INTERVAL 🔒

NETCHECK_REPORT_TIMEOUT 🔒

Maximum duration to wait for a netcheck report.

Statics

ADDR_COUNTER 🔒

Counter to always generate unique addresses for QuicMappedAddr.

Functions

bind 🔒

Initial connection setup.

disco_message_sent 🔒

endpoint_sets_equal 🔒

Reports whether x and y represent the same set of endpoints. The order doesn’t matter.

new_re_stun_timer 🔒

split_packets 🔒

Split a number of transmits into individual packets.

Type Aliases

RelayContents 🔒

Contents of a relay message. Use a SmallVec to avoid allocations for the very common case of a single packet.

RelayRecvResult 🔒