Module iroh_net::magicsock::node_map

Re-exports

pub use node_state::ConnectionType;

pub use node_state::ControlMsg;

pub use node_state::DirectAddrInfo;

pub use node_state::NodeInfo;

Modules

best_addr 🔒

The BestAddr is the currently active best address for UDP sends.

node_state 🔒

Structs

ConnectionTypeStream

Stream returning ConnectionTypes

IpPort

An (Ip, Port) pair.

NodeMap 🔒

Map of the NodeState information for all the known nodes.

NodeMapInner 🔒

Enums

NodeStateKey 🔒

Identifier to look up a NodeState in the NodeMap.

Source 🔒

Source for a new node.

Constants

MAX_INACTIVE_NODES 🔒

Number of nodes that are inactive for which we keep info about. This limit is enforced periodically via NodeMap::prune_inactive.