Module iroh_net::magicsock::node_map::node_state

Structs

DirectAddrInfo

Information about a direct address.

NodeInfo

Details about an iroh node which is known to this node.

NodeState 🔒

An iroh node, which we can have connections with.

Options 🔒

Options for creating a new NodeState.

PathState 🔒

State about a particular path to another NodeState.

PingHandled

Indicating an NodeState has handled a ping.

PongReply 🔒

RelayUrlInfo

Information about a relay URL.

SendPing 🔒

SentPing 🔒

Enums

ConnectionType

The type of connection we have to the endpoint.

ControlMsg

The type of control message we have received.

DiscoPingPurpose

The reason why a discovery ping message was sent.

PingAction 🔒

PingRole

SendCallMeMaybe 🔒

Whether to send a call-me-maybe message after sending pings to all known paths.

Constants

DISCO_PING_INTERVAL 🔒

The minimum time between pings to an endpoint. (Except in the case of CallMeMaybe frames resetting the counter, as the first pings likely didn’t through the firewall)

GOOD_ENOUGH_LATENCY 🔒

The latency at or under which we don’t try to upgrade to a better path.

LAST_ALIVE_PRUNE_DURATION 🔒

How long since an endpoint path was last alive before it might be pruned.

MAX_INACTIVE_DIRECT_ADDRESSES 🔒

Number of addresses that are not active that we keep around per node.

PING_TIMEOUT_DURATION 🔒

How long we wait for a pong reply before assuming it’s never coming.

SESSION_ACTIVE_TIMEOUT 🔒

How long since the last activity we try to keep an established endpoint peering alive. It’s also the idle time at which we stop doing STUN queries to keep NAT mappings alive.

STAYIN_ALIVE_MIN_ELAPSED 🔒

How long until we send a stayin alive ping

UPGRADE_INTERVAL 🔒

How often we try to upgrade to a better patheven if we have some non-relay route that works.

Functions

summarize_node_paths 🔒