Struct iroh_net::magicsock::node_map::node_state::NodeState

pub(super) struct NodeState {
    id: usize,
    quic_mapped_addr: QuicMappedAddr,
    node_id: NodeId,
    last_full_ping: Option<Instant>,
    relay_url: Option<(RelayUrl, PathState)>,
    best_addr: BestAddr,
    direct_addr_state: BTreeMap<IpPort, PathState>,
    sent_pings: HashMap<TransactionId, SentPing>,
    last_used: Option<Instant>,
    last_call_me_maybe: Option<Instant>,
    conn_type: Watchable<ConnectionType>,
}

An iroh node, which we can have connections with.

The whole point of the magicsock is that we can have multiple paths to a particular node. One of these paths is via the endpoint’s home relay node but as we establish a connection we’ll hopefully discover more direct paths.

Fields
id: usize

The ID used as index in the NodeMap.

quic_mapped_addr: QuicMappedAddr

The UDP address used on the QUIC-layer to address this node.

node_id: NodeId

The global identifier for this endpoint.

last_full_ping: Option<Instant>

The last time we pinged all endpoints.

relay_url: Option<(RelayUrl, PathState)>

The url of relay node that we can relay over to communicate.

The fallback/bootstrap path, if non-zero (non-zero for well-behaved clients).

best_addr: BestAddr

Best non-relay path, i.e. a UDP address.

direct_addr_state: BTreeMap<IpPort, PathState>

State for each of this node’s direct paths.

sent_pings: HashMap<TransactionId, SentPing>last_used: Option<Instant>

Last time this node was used.

A node is marked as in use when an endpoint to contact them is requested or if UDP activity is registered.

last_call_me_maybe: Option<Instant>

Last time we sent a call-me-maybe.

When we do not have a direct connection and we try to send some data, we will try to do a full ping + call-me-maybe. Usually each side only needs to send one call-me-maybe to the other for holes to be punched in both directions however. So we only try and send one per HEARTBEAT_INTERVAL. Each HEARTBEAT_INTERVAL the NodeState::stayin_alive function is called, which will trigger new call-me-maybe messages as backup.

conn_type: Watchable<ConnectionType>

The type of connection we have to the node, either direct, relay, mixed, or none.

Implementations

impl NodeState

pub(super) fn new(id: usize, options: Options) -> Self

pub(super) fn public_key(&self) -> &PublicKey

pub(super) fn quic_mapped_addr(&self) -> &QuicMappedAddr

pub(super) fn id(&self) -> usize

pub(super) fn conn_type(&self) -> ConnectionType

pub(super) fn conn_type_stream(&self) -> WatcherStream<ConnectionType>

pub(super) fn info(&self, now: Instant) -> NodeInfo

Returns info about this endpoint

pub(super) fn relay_url(&self) -> Option<RelayUrl>

Returns the relay url of this endpoint

fn addr_for_send( &mut self, now: &Instant ) -> (Option<SocketAddr>, Option<RelayUrl>)

Returns the address(es) that should be used for sending the next packet.

This may return to send on one, both or no paths.

pub(super) fn remove_direct_addr( &mut self, ip_port: &IpPort, reason: ClearReason )

Removes a direct address for this node.

If this is also the best address, it will be cleared as well.

fn assign_best_addr_from_candidates_if_empty(&mut self)

Fixup best_adrr from candidates.

If somehow we end up in a state where we failed to set a best_addr, while we do have valid candidates, this will chose a candidate and set best_addr again. Most likely this is a bug elsewhere though.

fn want_call_me_maybe(&self, now: &Instant) -> bool

Whether we need to send another call-me-maybe to the endpoint.

Basically we need to send a call-me-maybe if we need to find a better path. Maybe we only have a relay path, or our path is expired.

When a call-me-maybe message is sent we also need to send pings to all known paths of the endpoint. The NodeState::send_call_me_maybe function takes care of this.

pub(super) fn ping_timeout(&mut self, txid: TransactionId)

Cleanup the expired ping for the passed in txid.

fn start_ping( &self, dst: SendAddr, purpose: DiscoPingPurpose ) -> Option<SendPing>

pub(super) fn ping_sent( &mut self, to: SendAddr, tx_id: TransactionId, purpose: DiscoPingPurpose, sender: Sender<ActorMessage> )

Record the fact that a ping has been sent out.

fn send_call_me_maybe( &mut self, now: Instant, always: SendCallMeMaybe ) -> Vec<PingAction>

Send a DISCO call-me-maybe message to the peer.

This takes care of sending the needed pings beforehand. This ensures that we open our firewall’s port so that when the receiver sends us DISCO pings in response to our call-me-maybe they will reach us and the other side establishes a direct connection upon our subsequent pong response.

For SendCallMeMaybe::IfNoRecent, no paths will be pinged if there already was a recent call-me-maybe sent.

The caller is responsible for sending the messages.

fn send_pings(&mut self, now: Instant) -> Vec<PingAction>

Send DISCO Pings to all the paths of this node.

Any paths to the node which have not been recently pinged will be sent a disco ping.

The caller is responsible for sending the messages.

pub(super) fn update_from_node_addr(&mut self, n: &AddrInfo)

pub(super) fn reset(&mut self)

Clears all the endpoint’s p2p state, reverting it to a relay-only endpoint.

pub(super) fn handle_ping( &mut self, path: SendAddr, tx_id: TransactionId ) -> PingHandled

Handle a received Disco Ping.

Ensures the paths the ping was received on is a known path for this endpoint.

If there is no best_addr for this endpoint yet, sends a ping itself to try and establish one.

This is called once we’ve already verified that we got a valid discovery message from self via ep.

pub(super) fn prune_direct_addresses(&mut self)

Prune inactive paths.

This trims the list of inactive paths for an endpoint. At most MAX_INACTIVE_DIRECT_ADDRESSES are kept.

pub(super) fn note_connectivity_change(&mut self)

Called when connectivity changes enough that we should question our earlier assumptions about which paths work.

pub(super) fn handle_pong( &mut self, m: &Pong, src: SendAddr ) -> Option<(SocketAddr, PublicKey)>

Handles a Pong message (a reply to an earlier ping).

It reports the address and key that should be inserted for the endpoint if any.

pub(super) fn handle_call_me_maybe(&mut self, m: CallMeMaybe) -> Vec<PingAction>

Handles a DISCO CallMeMaybe discovery message.

The contract for use of this message is that the node has already pinged to us via UDP, so their stateful firewall should be open. Now we can Ping back and make it through.

However if the remote side has no direct path information to us, they would not have had any IpPorts to send pings to and our pings might end up blocked. But at least open the firewalls on our side, giving the other side another change of making it through when it pings in response.

pub(super) fn receive_udp(&mut self, addr: IpPort, now: Instant)

Marks this endpoint as having received a UDP payload message.

pub(super) fn receive_relay( &mut self, url: &RelayUrl, src: NodeId, now: Instant )

pub(super) fn last_ping(&self, addr: &SendAddr) -> Option<Instant>

pub(super) fn is_active(&self, now: &Instant) -> bool

Checks if this Endpoint is currently actively being used.

pub(super) fn stayin_alive(&mut self) -> Vec<PingAction>

Send a heartbeat to the node to keep the connection alive, or trigger a full ping if necessary.

pub(crate) fn get_send_addrs( &mut self ) -> (Option<SocketAddr>, Option<RelayUrl>, Vec<PingAction>)

Returns the addresses on which a payload should be sent right now.

This is in the hot path of .poll_send().

pub(super) fn direct_addresses(&self) -> impl Iterator<Item = IpPort> + '_

Get the direct addresses for this endpoint.

pub(super) fn last_used(&self) -> Option<Instant>