Struct iroh_net::magicsock::Actor

struct Actor {

    msock: Arc<MagicSock>,
    msg_receiver: Receiver<ActorMessage>,
    msg_sender: Sender<ActorMessage>,
    relay_actor_sender: Sender<RelayActorMessage>,
    relay_actor_cancel_token: CancellationToken,
    relay_recv_sender: Sender<Result<(PublicKey, RecvMeta, Bytes), Error>>,
    periodic_re_stun_timer: Interval,
    net_info_last: Option<NetInfo>,
    pconn4: UdpConn,
    pconn6: Option<UdpConn>,
    port_mapper: Client,
    no_v4_send: bool,
    net_checker: Client,
    network_monitor: Monitor,

}

Fields
msock: Arc<MagicSock>msg_receiver: Receiver<ActorMessage>msg_sender: Sender<ActorMessage>relay_actor_sender: Sender<RelayActorMessage>relay_actor_cancel_token: CancellationTokenrelay_recv_sender: Sender<Result<(PublicKey, RecvMeta, Bytes), Error>>

Channel to send received relay messages on, for processing.

periodic_re_stun_timer: Interval

When set, is an AfterFunc timer that will call MagicSock::do_periodic_stun.

net_info_last: Option<NetInfo>

The NetInfo provided in the last call to net_info_func. It’s used to deduplicate calls to netInfoFunc.

pconn4: UdpConnpconn6: Option<UdpConn>port_mapper: Client

The NAT-PMP/PCP/UPnP prober/client, for requesting port mappings from NAT devices.

no_v4_send: bool

Whether IPv4 UDP is known to be unable to transmit at all. This could happen if the socket is in an invalid state (as can happen on darwin after a network link status change).

net_checker: Client

The prober that discovers local network conditions, including the closest relay relay and NAT mappings.

network_monitor: Monitor
Implementations

impl Actor

async fn run(self) -> Result<()>

async fn handle_network_change(&mut self, is_major: bool)

async fn handle_ping_actions(&mut self, msgs: Vec<PingAction>)

async fn handle_actor_message(&mut self, msg: ActorMessage) -> bool

Processes an incoming actor message.

Returns true if it was a shutdown.

fn normalized_local_addr(&self) -> Result<SocketAddr>

fn local_addr(&self) -> (Result<SocketAddr>, Option<Result<SocketAddr>>)

fn process_relay_read_result( &mut self, dm: RelayReadResult ) -> Vec<Result<(PublicKey, RecvMeta, Bytes), Error>>

async fn update_direct_addrs(&mut self, why: &'static str)

Refreshes knowledge about our direct addresses.

In other words, this triggers a netcheck run.

Note that invoking this is managed by the DirectAddrUpdateState and this should never be invoked directly. Some day this will be refactored to not allow this easy mistake to be made.

async fn store_direct_addr_update(&mut self, nr: Option<Arc<Report>>)

Stores the results of a successful direct addr update.

fn finalize_direct_addrs_update(&mut self, why: &'static str)

Called when a direct addr update is done, no matter if it was successful or not.

async fn set_net_info_have_port_map(&mut self)

Updates NetInfo.HavePortMap to true.

async fn call_net_info_callback(&mut self, ni: NetInfo)

async fn update_net_info(&mut self, why: &'static str)

Calls netcheck.

Note that invoking this is managed by DirectAddrUpdateState via Actor::update_direct_addrs and this should never be invoked directly. Some day this will be refactored to not allow this easy mistake to be made.

async fn handle_netcheck_report(&mut self, report: Option<Arc<Report>>)

fn set_nearest_relay(&mut self, relay_url: Option<RelayUrl>) -> bool

fn pick_relay_fallback(&self) -> Option<RelayUrl>

Returns a deterministic relay node to connect to. This is only used if netcheck couldn’t find the nearest one, for instance, if UDP is blocked and thus STUN latency checks aren’t working.

If no the RelayMap is empty, returns 0.

fn reset_endpoint_states(&mut self)

Resets the preferred address for all nodes. This is called when connectivity changes enough that we no longer trust the old routes.

async fn close_stale_relay_connections(&self)

Tells the relay actor to close stale relay connections.

The relay connections who’s local endpoints no longer exist after a network change will error out soon enough. Closing them eagerly speeds this up however and allows re-establishing a relay connection faster.

fn send_relay_actor(&self, msg: RelayActorMessage)

fn handle_relay_disco_message( &mut self, msg: &[u8], url: &RelayUrl, relay_node_src: PublicKey ) -> bool