Struct iroh_net::magicsock::MagicSock

pub(crate) struct MagicSock {

    actor_sender: Sender<ActorMessage>,
    relay_actor_sender: Sender<RelayActorMessage>,
    me: String,
    proxy_url: Option<Url>,
    relay_recv_receiver: Receiver<Result<(PublicKey, RecvMeta, Bytes), Error>>,
    network_recv_wakers: Mutex<Option<Waker>>,
    network_send_wakers: Mutex<Option<Waker>>,
    dns_resolver: DnsResolver,
    secret_key: SecretKey,
    local_addrs: RwLock<(SocketAddr, Option<SocketAddr>)>,
    port: AtomicU16,
    closing: AtomicBool,
    closed: AtomicBool,
    ipv6_reported: Arc<AtomicBool>,
    relay_map: RelayMap,
    my_relay: Watchable<Option<RelayUrl>>,
    node_map: NodeMap,
    pconn4: UdpConn,
    pconn6: Option<UdpConn>,
    net_checker: Client,
    disco_secrets: DiscoSecrets,
    udp_state: UdpState,
    send_buffer: Mutex<Vec<Transmit>>,
    udp_disco_sender: Sender<(SocketAddr, PublicKey, Message)>,
    discovery: Option<Box<dyn Discovery>>,
    direct_addrs: Watchable<DiscoveredDirectAddrs>,
    pending_call_me_maybes: Mutex<HashMap<PublicKey, RelayUrl>>,
    direct_addr_update_state: DirectAddrUpdateState,
    insecure_skip_relay_cert_verify: bool,

}

Iroh connectivity layer.

This is responsible for routing packets to nodes based on node IDs, it will initially route packets via a relay and transparently try and establish a node-to-node connection and upgrade to it. It will also keep looking for better connections as the network details of both nodes change.

It is usually only necessary to use a single MagicSock instance in an application, it means any QUIC endpoints on top will be sharing as much information about nodes as possible.

Fields
actor_sender: Sender<ActorMessage>relay_actor_sender: Sender<RelayActorMessage>me: String

String representation of the node_id of this node.

proxy_url: Option<Url>

Proxy

relay_recv_receiver: Receiver<Result<(PublicKey, RecvMeta, Bytes), Error>>

Used for receiving relay messages.

network_recv_wakers: Mutex<Option<Waker>>

Stores wakers, to be called when relay_recv_ch receives new data.

network_send_wakers: Mutex<Option<Waker>>dns_resolver: DnsResolver

The DNS resolver to be used in this magicsock.

secret_key: SecretKey

Key for this node.

local_addrs: RwLock<(SocketAddr, Option<SocketAddr>)>

Cached version of the Ipv4 and Ipv6 addrs of the current connection.

port: AtomicU16

Preferred port from Options::port; 0 means auto.

closing: AtomicBool

Close is in progress (or done)

closed: AtomicBool

Close was called.

ipv6_reported: Arc<AtomicBool>

If the last netcheck report, reports IPv6 to be available.

relay_map: RelayMap

None (or zero nodes) means relay is disabled.

my_relay: Watchable<Option<RelayUrl>>

Nearest relay node ID; 0 means none/unknown.

node_map: NodeMap

Tracks the networkmap node entity for each node discovery key.

pconn4: UdpConn

UDP IPv4 socket

pconn6: Option<UdpConn>

UDP IPv6 socket

net_checker: Client

Netcheck client

disco_secrets: DiscoSecrets

The state for an active DiscoKey.

udp_state: UdpStatesend_buffer: Mutex<Vec<Transmit>>

Send buffer used in poll_send_udp

udp_disco_sender: Sender<(SocketAddr, PublicKey, Message)>

UDP disco (ping) queue

discovery: Option<Box<dyn Discovery>>

Optional discovery service

direct_addrs: Watchable<DiscoveredDirectAddrs>

Our discovered direct addresses.

pending_call_me_maybes: Mutex<HashMap<PublicKey, RelayUrl>>

List of CallMeMaybe disco messages that should be sent out after the next endpoint update completes

direct_addr_update_state: DirectAddrUpdateState

Indicates the direct addr update state.

insecure_skip_relay_cert_verify: bool

Skip verification of SSL certificates from relay servers

May only be used in tests.

Implementations

impl MagicSock

pub(crate) async fn spawn(opts: Options) -> Result<Handle>

Creates a magic MagicSock listening on Options::port.

pub(crate) fn my_relay(&self) -> Option<RelayUrl>

Returns the relay node we are connected to, that has the best latency.

If None, then we are not connected to any relay nodes.

pub(crate) fn proxy_url(&self) -> Option<&Url>

Get the current proxy configuration.

fn set_my_relay(&self, my_relay: Option<RelayUrl>) -> Option<RelayUrl>

Sets the relay node with the best latency.

If we are not connected to any relay nodes, set this to None.

fn is_closing(&self) -> bool

fn is_closed(&self) -> bool

fn public_key(&self) -> PublicKey

pub(crate) fn local_addr(&self) -> (SocketAddr, Option<SocketAddr>)

Get the cached version of the Ipv4 and Ipv6 addrs of the current connection.

pub(crate) fn has_send_address(&self, node_key: PublicKey) -> bool

Returns true if we have at least one candidate address where we can send packets to.

pub(crate) fn connection_infos(&self) -> Vec<ConnectionInfo>

Retrieve connection information about nodes in the network.

pub(crate) fn connection_info(&self, node_id: NodeId) -> Option<ConnectionInfo>

Retrieve connection information about a node in the network.

pub(crate) fn direct_addresses(&self) -> DirectAddrsStream

Returns the direct addresses as a stream.

The MagicSock continuously monitors the direct addresses, the network addresses it might be able to be contacted on, for changes. Whenever changes are detected this stream will yield a new list of addresses.

Upon the first creation on the MagicSock it may not yet have completed a first direct addresses discovery, in this case the first item of the stream will not be immediately available. Once this first set of direct addresses are discovered the stream will always return the first set of addresses immediately, which are the most recently discovered addresses.

To get the current direct addresses, drop the stream after the first item was received.

pub(crate) fn watch_home_relay(&self) -> impl Stream<Item = RelayUrl>

Watch for changes to the home relay.

Note that this can be used to wait for the initial home relay to be known. If the home relay is known at this point, it will be the first item in the stream.

pub(crate) fn conn_type_stream( &self, node_id: NodeId ) -> Result<ConnectionTypeStream>

Returns a stream that reports the ConnectionType we have to the given node_id.

The NodeMap continuously monitors the node_id’s endpoint for ConnectionType changes, and sends the latest ConnectionType on the stream.

The current ConnectionType will the the initial entry on the stream.

Errors

Will return an error if there is no address information known about the given node_id.

pub(crate) fn get_mapping_addr(&self, node_id: NodeId) -> Option<SocketAddr>

Returns the SocketAddr which can be used by the QUIC layer to dial this node.

Note this is a user-facing API and does not wrap the SocketAddr in a QuicMappedAddr as we do internally.

pub fn add_node_addr(&self, addr: NodeAddr, source: Source) -> Result<()>

Add addresses for a node to the magic socket’s addresbook.

pub(crate) fn update_direct_addresses(&self, eps: Vec<DirectAddr>)

Updates our direct addresses.

On a successful update, our address is published to discovery.

pub(crate) fn dns_resolver(&self) -> &DnsResolver

Get a reference to the DNS resolver used in this MagicSock.

pub(crate) fn discovery(&self) -> Option<&dyn Discovery>

Reference to optional discovery service

pub(crate) async fn network_change(&self)

Call to notify the system of potential network changes.

fn normalized_local_addr(&self) -> Result<SocketAddr>

fn poll_send( &self, cx: &mut Context<'_>, transmits: &[Transmit] ) -> Poll<Result<usize>>

fn poll_send_udp( &self, addr: SocketAddr, transmits: &[Transmit], cx: &mut Context<'_> ) -> Poll<Result<usize>>

fn conn_for_addr(&self, addr: SocketAddr) -> Result<&UdpConn>

fn poll_recv( &self, cx: &mut Context<'_>, bufs: &mut [IoSliceMut<'_>], metas: &mut [RecvMeta] ) -> Poll<Result<usize>>

NOTE: Receiving on a Self::closed socket will return Poll::Pending indefinitely.

fn poll_recv_relay( &self, cx: &mut Context<'_>, bufs: &mut [IoSliceMut<'_>], metas: &mut [RecvMeta] ) -> Poll<Result<usize>>

fn handle_disco_message( &self, sender: PublicKey, sealed_box: &[u8], src: DiscoMessageSource )

Handles a discovery message.

fn handle_ping(&self, dm: Ping, sender: NodeId, src: DiscoMessageSource)

Handle a ping message.

fn encode_disco_message(&self, dst_key: PublicKey, msg: &Message) -> Bytes

fn send_ping_queued(&self, ping: SendPing)

fn poll_send_ping( &self, ping: &SendPing, cx: &mut Context<'_> ) -> Poll<Result<()>>

fn send_disco_message_queued( &self, dst: SendAddr, dst_key: PublicKey, msg: Message ) -> bool

Send a disco message. UDP messages will be queued.

If dst is SendAddr::Relay, the message will be pushed into the relay client channel. If dst is SendAddr::Udp, the message will be pushed into the udp disco send channel.

Returns true if the channel had capacity for the message, and false if the message was dropped.

fn poll_send_disco_message( &self, dst: SendAddr, dst_key: PublicKey, msg: Message, cx: &mut Context<'_> ) -> Poll<Result<()>>

Send a disco message. UDP messages will be polled to send directly on the UDP socket.

fn send_disco_message_relay( &self, url: &RelayUrl, dst_key: PublicKey, msg: Message ) -> bool

async fn send_disco_message_udp( &self, dst: SocketAddr, dst_key: PublicKey, msg: &Message ) -> Result<bool>

fn poll_send_disco_message_udp( &self, dst: SocketAddr, dst_key: PublicKey, msg: &Message, cx: &mut Context<'_> ) -> Poll<Result<bool>>

fn poll_handle_ping_actions( &self, cx: &mut Context<'_>, msgs: &mut Vec<PingAction> ) -> Poll<Result<()>>

fn poll_handle_ping_action( &self, cx: &mut Context<'_>, msg: &PingAction ) -> Poll<Result<()>>

fn poll_send_relay( &self, url: &RelayUrl, node: PublicKey, contents: SmallVec<[Bytes; 1]> ) -> Poll<bool>

fn send_queued_call_me_maybes(&self)

fn send_or_queue_call_me_maybe(&self, url: &RelayUrl, dst_node: NodeId)

fn re_stun(&self, why: &'static str)

Triggers an address discovery. The provided why string is for debug logging only.

fn publish_my_addr(&self)

Publishes our address to a discovery service, if configured.

Called whenever our addresses or home relay node changes.