Struct iroh_net::net::UdpSocket

pub struct UdpSocket(/* private fields */);

Wrapper around a tokio UDP socket that handles the fact that on drop libc::close can block for UDP sockets.

Implementations

impl UdpSocket

pub fn bind_v4(port: u16) -> Result<Self>

Bind only Ipv4 on any interface.

pub fn bind_v6(port: u16) -> Result<Self>

Bind only Ipv6 on any interface.

pub fn bind_local_v4(port: u16) -> Result<Self>

Bind only Ipv4 on localhost.

pub fn bind_local_v6(port: u16) -> Result<Self>

Bind only Ipv6 on localhost.

pub fn bind_local(network: IpFamily, port: u16) -> Result<Self>

Bind to the given port only on localhost.

pub fn bind(network: IpFamily, port: u16) -> Result<Self>

Bind to the given port and listen on all interfaces.

pub fn bind_full(addr: impl Into<SocketAddr>) -> Result<Self>

Bind to any provided SocketAddr.