Module iroh_net::netcheck

Checks the network conditions from the current host.

Netcheck is responsible for finding out the network conditions of the current host, like whether it is connected to the internet via IPv4 and/or IPv6, what the NAT situation is etc.

Based on https://github.com/tailscale/tailscale/blob/main/net/netcheck/netcheck.go

Modules

metrics 🔒

reportgen 🔒

The reportgen actor is responsible for generating a single netcheck report.

Structs

Actor 🔒

The netcheck actor.

Addr 🔒

Sender to the Actor.

Client

Client to run netchecks.

Inflight 🔒

Metrics

Enum of metrics for the module

RelayLatencies

Latencies per relay node.

Report

A netcheck report.

ReportRun 🔒

State the netcheck actor needs for an in-progress report generation.

Reports 🔒

Enums

Message 🔒

Messages to send to the Actor.

Constants

DEFAULT_MAX_LATENCY 🔒

The maximum latency of all nodes, if none are found yet.

FULL_REPORT_INTERVAL 🔒

Functions

bind_local_stun_socket 🔒

Attempts to bind a local socket to send STUN packets from.

os_has_ipv6 🔒

Test if IPv6 works at all, or if it’s been hard disabled at the OS level.

recv_stun_once 🔒

Receive STUN response from a UDP socket, pass it to the actor.