Module iroh_net::netcheck::reportgen::probes

The relay probes.

All the probes try and establish the latency to the relay servers. Preferably the STUN probes work and we also learn about our public IP addresses and ports. But fallback probes for HTTPS and ICMP exist as well.

Structs

ProbePlan 🔒

A probe plan.

ProbeSet 🔒

A probe set is a sequence of similar Probes with delays between them.

Enums

Probe 🔒

ProbeProto 🔒

The protocol used to time a node’s latency.

Constants

ACTIVE_RETRANSMIT_EXTRA_DELAY 🔒

The extra time to add to retransmits if a previous report exists.

DEFAULT_ACTIVE_RETRANSMIT_DELAY 🔒

The retransmit interval used when a previous report exists but is missing latency.

DEFAULT_INITIAL_RETRANSMIT 🔒

The retransmit interval used when netcheck first runs.

NUM_INCREMENTAL_RELAYS 🔒

The number of fastest relays to periodically re-query during incremental netcheck reports. (During a full report, all relay servers are scanned.)

Functions

sort_relays 🔒

Sorts the nodes in the RelayMap from fastest to slowest.