Struct iroh_net::netcheck::reportgen::Actor

struct Actor {
    msg_tx: Sender<Message>,
    msg_rx: Receiver<Message>,
    netcheck: Addr,
    last_report: Option<Arc<Report>>,
    port_mapper: Option<Client>,
    relay_map: RelayMap,
    stun_sock4: Option<Arc<UdpSocket>>,
    stun_sock6: Option<Arc<UdpSocket>>,
    report: Report,
    hairpin_actor: Client,
    outstanding_tasks: OutstandingTasks,
    dns_resolver: DnsResolver,
}

The reportstate actor.

This actor starts, generates a single report and exits.

Fields
msg_tx: Sender<Message>

The sender of the message channel, so we can give out Addr.

msg_rx: Receiver<Message>

The receiver of the message channel.

netcheck: Addr

The address of the netcheck actor.

last_report: Option<Arc<Report>>

The previous report, if it exists.

port_mapper: Option<Client>

The portmapper client, if there is one.

relay_map: RelayMap

The relay configuration.

stun_sock4: Option<Arc<UdpSocket>>

Socket to send IPv4 STUN requests from.

stun_sock6: Option<Arc<UdpSocket>>

Socket so send IPv6 STUN requests from.

report: Report

The report being built.

hairpin_actor: Client

The hairping actor.

outstanding_tasks: OutstandingTasks

Which tasks the Actor is still waiting on.

This is essentially the summary of all the work the Actor is doing.

dns_resolver: DnsResolver

The DNS resolver to use for probes that need to resolve DNS records.

Implementations

impl Actor

fn addr(&self) -> Addr

async fn run(&mut self)

async fn run_inner(&mut self) -> Result<()>

Runs the main reportgen actor logic.

This actor runs by:

Creates a hairpin actor.

Creates a captive portal future.

Creates ProbeSet futures in a group of futures.

Runs a main loop:

Drives all the above futures.

Receives actor messages (sent by those futures).

Updates the report, cancels unneeded futures.

Sends the report to the netcheck actor.

fn handle_message(&mut self, msg: Message)

Handles an actor message.

Returns true if all the probes need to be aborted.

fn handle_probe_report(&mut self, probe_report: ProbeReport)

fn probe_would_help(&mut self, probe: Probe, relay_node: Arc<RelayNode>) -> bool

Whether running this probe would still improve our report.

fn handle_abort_probes(&mut self)

Stops further probes.

This makes sure that no further probes are run and also cancels the captive portal and portmapper tasks if there were successful probes. Be sure to only handle this after all the required ProbeReports have been processed.

fn prepare_portmapper_task( &mut self ) -> MaybeFuture<Pin<Box<impl Future<Output = Option<ProbeOutput>>>>>

Creates the future which will perform the portmapper task.

The returned future will run the portmapper, if enabled, resolving to it’s result.

fn prepare_captive_portal_task( &mut self ) -> MaybeFuture<Pin<Box<impl Future<Output = Option<bool>>>>>

Creates the future which will perform the captive portal check.

async fn spawn_probes_task(&mut self) -> Result<JoinSet<Result<ProbeReport>>>

Prepares the future which will run all the probes as per generated ProbePlan.

Probes operate like the following:

A future is created for each probe in all probe sets.

All probes in a set are grouped in [JoinSet].

All those probe sets are grouped in one overall [JoinSet].

This future is polled by the main actor loop to make progress.

Once a probe future is polled:

Many probes start with a delay, they sleep during this time.

When a probe starts it first asks the reportgen Actor if it is still useful to run. If not it aborts the entire probe set.

When a probe finishes, its ProbeReport is yielded to the reportgen actor.

Probes get aborted in several ways:

A running it can fail and abort the entire probe set if it deems the failure permanent. Probes in a probe set are essentially retries.

Once there are ProbeReports from enough nodes, all remaining probes are aborted. That is, the main actor loop stops polling them.