Struct iroh_net::netcheck::Actor

struct Actor {
    receiver: Receiver<Message>,
    sender: Sender<Message>,
    reports: Reports,
    port_mapper: Option<Client>,
    in_flight_stun_requests: HashMap<TransactionId, Inflight>,
    current_report_run: Option<ReportRun>,
    dns_resolver: DnsResolver,
}

The netcheck actor.

This actor runs for the entire duration there’s a Client connected.

Fields
receiver: Receiver<Message>

Actor messages channel.

If there are no more senders the actor stops.

sender: Sender<Message>

The sender side of the messages channel.

This allows creating new Addrs from the actor.

reports: Reports

A collection of previously generated reports.

Sometimes it is useful to look at past reports to decide what to do.

port_mapper: Option<Client>

The port mapper client, if those are requested.

The port mapper is responsible for talking to routers via UPnP and the like to try and open ports.

in_flight_stun_requests: HashMap<TransactionId, Inflight>

Information about the currently in-flight STUN requests.

This is used to complete the STUN probe when receiving STUN packets.

current_report_run: Option<ReportRun>

The reportgen actor currently generating a report.

dns_resolver: DnsResolver

The DNS resolver to use for probes that need to perform DNS lookups

Implementations

impl Actor

fn new(port_mapper: Option<Client>, dns_resolver: DnsResolver) -> Result<Self>

Creates a new actor.

This does not start the actor, see Actor::run for this. You should not normally create this directly but rather create a Client.

fn addr(&self) -> Addr

Returns the channel to send messages to the actor.

async fn run(&mut self)

Run the actor.

It will now run and handle messages. Once the connected Client (including all its clones) is dropped this will terminate.

fn handle_run_check( &mut self, relay_map: RelayMap, stun_sock_v4: Option<Arc<UdpSocket>>, stun_sock_v6: Option<Arc<UdpSocket>>, response_tx: Sender<Result<Arc<Report>>> )

Starts a check run as requested by the Message::RunCheck message.

If stun_sock_v4 or stun_sock_v6 are not provided this will bind the sockets itself. This is not ideal since really you want to send STUN probes from the sockets you will be using.

fn handle_report_ready(&mut self, report: Box<Report>)

fn handle_report_aborted(&mut self)

fn handle_stun_packet(&mut self, pkt: &[u8], src: SocketAddr)

Handles Message::StunPacket.

If there are currently no in-flight stun requests registered this is dropped, otherwise forwarded to the probe.

fn handle_in_flight_stun(&mut self, inflight: Inflight, response_tx: Sender<()>)

Handles Message::InFlightStun.

The in-flight request is added to Actor::in_flight_stun_requests so that Actor::handle_stun_packet can forward packets correctly.

response_tx is to signal the actor message has been handled.

fn finish_and_store_report(&mut self, report: Report) -> Arc<Report>

fn add_report_history_and_set_preferred_relay( &mut self, r: Report ) -> Arc<Report>

Adds r to the set of recent Reports and mutates r.preferred_relay to contain the best recent one. r is stored ref counted and a reference is returned.