Module iroh_net::portmapper

Port mapping client and service.

Modules

current_mapping 🔒

Holds the current mapping value and ensures that any change is reported accordingly.

mapping 🔒

A port mapping created with one of the supported protocols.

metrics 🔒

nat_pmp 🔒

Definitions and utilities to interact with a NAT-PMP server.

pcp 🔒

Definitions and utilities to interact with a PCP server.

upnp 🔒

Structs

Client

Port mapping client.

Config

Configures which port mapping protocols are enabled in the Service.

Metrics

Enum of metrics for the module

Probe 🔒

Port mapping protocol information obtained during a probe.

ProbeOutput

Output of a port mapping probe.

Service

A port mapping client.

Enums

Message 🔒

Constants

AVAILABILITY_TRUST_DURATION 🔒

If a port mapping service has been seen within the last AVAILABILITY_TRUST_DURATION it will not be probed again.

SERVICE_CHANNEL_CAPACITY 🔒

Capacity of the channel to communicate with the long-running service.

UNAVAILABILITY_TRUST_DURATION 🔒

If a port mapping service has not been seen within the last UNAVAILABILITY_TRUST_DURATION we allow trying a mapping using said protocol.

Functions

ip_and_gateway 🔒

Gets the local ip and gateway address for port mapping.

Type Aliases

ProbeResult 🔒