Module iroh_net::relay::client

based on tailscale/derp/derp_client.go

Structs

Client

A relay Client. Cheaply clonable. Call close to shutdown the write loop and read functionality.

ClientBuilder

The Builder returns a Client and a started ClientWriter run task.

ClientReceiver

ClientWriter ğŸ”’

Call ClientWriter::run to listen for messages to send to the client. Should be used by the Client

InnerClient

Enums

ClientWriterMessage ğŸ”’

The kinds of messages we can send to the super::server::Server

ConnReader ğŸ”’

ConnWriter ğŸ”’

ReceivedMessage

The type of message received by the Client from the super::server::Server.

Constants

CLIENT_RECV_TIMEOUT ğŸ”’

Functions

process_incoming_frame ğŸ”’

send_packet ğŸ”’

tung_wasm_to_io_err ğŸ”’