Module iroh_net::relay::codec

Structs

DerpCodec 🔒

Enums

Frame 🔒

FrameType 🔒

The one byte frame type at the beginning of the frame header. The second field is a big-endian u32 describing the length of the remaining frame (not including the initial 5 bytes)

Constants

HEADER_LEN 🔒

KEEP_ALIVE 🔒

MAGIC 🔒

The Relay magic number, sent in the FrameType::ClientInfo frame upon initial connection.

MAX_FRAME_SIZE 🔒

MAX_PACKET_SIZE

The maximum size of a packet sent over relay. (This only includes the data bytes visible to magicsock, not including its on-wire framing overhead)

NOT_PREFERRED 🔒

indicates this is NOT the client’s home node

PER_CLIENT_READ_QUEUE_DEPTH 🔒

PER_CLIENT_SEND_QUEUE_DEPTH 🔒

The number of packets buffered for sending per client

PREFERRED 🔒

Protocol flow:

PROTOCOL_VERSION 🔒

ProtocolVersion is bumped whenever there’s a wire-incompatible change.

SERVER_CHANNEL_SIZE 🔒

Functions

recv_client_key 🔒

Reads the FrameType::ClientInfo frame from the client (its proof of identity) upon it’s initial connection.

recv_frame 🔒

Receives the next frame and matches the frame type. If the correct type is found returns the content, otherwise an error.

send_client_key 🔒

Writes a FrameType::ClientInfo, including the client’s PublicKey, and the client’s ClientInfo, sealed using the server’s PublicKey.

write_frame 🔒

Writes complete frame, errors if it is unable to write within the given timeout. Ignores the timeout if None