Module iroh_net::relay::http::server

Structs

Handlers 🔒

Inner 🔒

RelayService 🔒

The hyper Service that servers the actual relay endpoints

Server

The Relay HTTP server.

ServerBuilder

Builder for the Relay HTTP Server.

ServerHandle

A handle for the Server.

ServerState 🔒

TlsConfig

Configuration to use for the TLS connection

Enums

RelayHandler 🔒

Action to take when a connection is made at the relay endpoint.`

TlsAcceptor

TLS Certificate Authority acceptor.

Functions

body_empty 🔒

Creates a new BytesBody with no content.

body_full 🔒

Creates a new BytesBody with given content.

downcast_upgrade 🔒

relay_connection_handler 🔒

The server HTTP handler to do HTTP upgrades.

Type Aliases

BytesBody 🔒

HyperError 🔒

HyperHandler 🔒

HyperResult 🔒