Module iroh_net::relay::iroh_relay

A full-fledged iroh-relay server.

This module provides an API to run a full fledged iroh-relay server. It is primarily used by the iroh-relay binary in this crate. It can be used to run a relay server in other locations however.

This code is fully written in a form of structured-concurrency: every spawned task is always attached to a handle and when the handle is dropped the tasks abort. So tasks can not outlive their handle. It is also always possible to await for completion of a task. Some tasks additionally have a method to do graceful shutdown.

Modules

metrics 🔒

Structs

CaptivePortalService 🔒

Limits

Rate limits.

RelayConfig

Configuration for the Relay HTTP and HTTPS server.

RelayHttpServerGuard 🔒

Horrible hack to make relay::http::Server behave somewhat.

Server

A running Relay + STUN server.

ServerConfig

Configuration for the full Relay & STUN server.

StunConfig

Configuration for the STUN server.

TlsConfig

TLS configuration for Relay server.

Enums

CertConfig

TLS certificate configuration.

Constants

INDEX 🔒

NOTFOUND 🔒

NO_CONTENT_CHALLENGE_HEADER 🔒

NO_CONTENT_RESPONSE_HEADER 🔒

RELAY_DISABLED 🔒

ROBOTS_TXT 🔒

TLS_HEADERS 🔒

Functions

body_empty 🔒

Creates a new BytesBody with no content.

handle_stun_request 🔒

Handles a single STUN request, doing all logging required.

is_challenge_char 🔒

probe_handler 🔒

HTTP latency queries

relay_disabled_handler 🔒

relay_supervisor 🔒

Supervisor for the relay server tasks.

robots_handler 🔒

root_handler 🔒

run_captive_portal_service 🔒

This is a future that never returns, drop it to cancel/abort.

serve_no_content_handler 🔒

For captive portal detection.

server_stun_listener 🔒

Runs a STUN server.

Type Aliases

BytesBody 🔒

HyperError 🔒

HyperResult 🔒